Atualização de esquema e implementação do serviceProvider, para garantir que o DB seja construído ao subir a aplicação.

+++ ECCUSBET Web/Models/Calculations/OrcamentoCalculation.cs
@@ -9,19 +9,19 @@ namespace ECCUSBET_Web.Models.Calculations
 {
     public class OrcamentoCalculation : OrcamentoEntity
     {
-        public OrcamentoCalculation(string servico, string equipamento, string material, double valorUnitario, int quantidade, double custoTotal) : base(servico, equipamento, material, valorUnitario, quantidade, custoTotal)
+        public OrcamentoCalculation(string servico, string equipamento, string material, decimal valorUnitario, int quantidade, decimal custoTotal) : base(servico, equipamento, material, valorUnitario, quantidade, custoTotal)
         {
         }
 
-        public double ValorTotaldeCadaItem()
+        public decimal ValorTotaldeCadaItem()
         {
            
             return Quantidade * ValorUnitario;
         }
 
-        public List<double> CustoTotaldaBacia()
+        public List<decimal> CustoTotaldaBacia()
         {
-            var lista = new List<double>();
+            var lista = new List<decimal>();
             lista.Add(ValorTotaldeCadaItem());
             return lista;
         }
